+++ ceammc/ext/src/proto/proto_midi.h
@@ -55,6 +55,7 @@ private:
 };
 
 class ProtoMidi : public BaseObject {
+protected:
     midi::MidiParser parser_;
     MidiQuaterFrame mqf_;
 
@@ -92,7 +93,7 @@ private:
 
     void byteStatus(uint8_t st, int chan) { floatTo(0, st | uint8_t(0x0F & chan)); }
     void byteData(uint8_t data) { floatTo(0, 0x7F & data); }
-    void msgTo(t_symbol* s, const Atom* a, size_t n) { anyTo(0, s, AtomListView(a, n)); }
+    void msgTo(t_symbol* s, const Atom* a, size_t n) { anyTo(1, s, AtomListView(a, n)); }
     void sendBytes3(int chan, uint8_t st, uint8_t data1, uint8_t data2);
 
     void handleTimecode(uint8_t data);